+++ CHANGES.rst
@@ -6,6 +6,13 @@ unreleased
 
 - Add jsdoc-conf.json, configuration file for jsdoc.
 
+- Intersection tool shows generic functionality for all 3-or-more-d layers
+
+Release 0.1
+-----------
+
+- Layer chooser is now a directive and has a background-image.
+
 - All clicks on the map result all data available to that location
 
 - ExtentAggregat is the default card displaying an extent summary of
@@ -26,9 +33,6 @@ unreleased
 - Events circles on map now don't increase with every redraw.
 
 - Events with start and end as line in timeline.
-
-Release 0.1
------------
 
 - Cumulative rain for spatial extent in timeline.